Remove UIContext watching responsibilities from TeamExplorerServiceHolder
This is now being handled by VSGitExt.
