+++ src/GitHub.TeamFoundation.14/Base/TeamExplorerServiceHolder.cs
@@ -1,8 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.Composition;
-using System.Diagnostics;
-using System.Globalization;
 using System.Linq;
 using System.Threading;
 using GitHub.Extensions;
@@ -10,8 +8,6 @@ using GitHub.Logging;
 using GitHub.Models;
 using GitHub.Services;
 using Microsoft.TeamFoundation.Controls;
-using Microsoft.VisualStudio.Shell;
-using Microsoft.VisualStudio.TeamFoundation.Git.Extensibility;
 using Serilog;
 
 namespace GitHub.VisualStudio.Base
@@ -26,9 +22,7 @@ namespace GitHub.VisualStudio.Base
         bool activeRepoNotified = false;
 
         IServiceProvider serviceProvider;
-        IVSGitExt gitService;
-        IVSUIContext gitUIContext;
-        IVSUIContextFactory uiContextFactory;
+        readonly IVSGitExt gitService;
 
         // ActiveRepositories PropertyChanged event comes in on a non-main thread
         readonly SynchronizationContext syncContext;
@@ -43,9 +37,13 @@ namespace GitHub.VisualStudio.Base
         [ImportingConstructor]
         public TeamExplorerServiceHolder(IVSGitExt gitService)
         {
-            this.GitService = gitService;
+            this.gitService = gitService;
             syncContext = SynchronizationContext.Current;
+
+            UpdateActiveRepo();
+            gitService.ActiveRepositoriesChanged += UpdateActiveRepo;
         }
+
 
         // set by the sections when they get initialized
         public IServiceProvider ServiceProvider
@@ -59,8 +57,6 @@ namespace GitHub.VisualStudio.Base
                 serviceProvider = value;
                 if (serviceProvider == null)
                     return;
-                GitUIContext = GitUIContext ?? UIContextFactory.GetUIContext(new Guid(Guids.GitSccProviderId));
-                UIContextChanged(GitUIContext?.IsActive ?? false, false);
             }
         }
 
@@ -133,12 +129,6 @@ namespace GitHub.VisualStudio.Base
             ServiceProvider = null;
         }
 
-        public void Refresh()
-        {
-            GitUIContext = GitUIContext ?? UIContextFactory.GetUIContext(new Guid(Guids.GitSccProviderId));
-            UIContextChanged(GitUIContext?.IsActive ?? false, true);
-        }
-
         void NotifyActiveRepo()
         {
             lock (activeRepoHandlers)
@@ -147,54 +137,6 @@ namespace GitHub.VisualStudio.Base
                 foreach (var handler in activeRepoHandlers.Values)
                     handler(activeRepo);
             }
-        }
-
-        void UIContextChanged(object sender, IVSUIContextChangedEventArgs e)
-        {
-            Guard.ArgumentNotNull(e, nameof(e));
-
-            ActiveRepo = null;
-            UIContextChanged(e.Activated, false);
-        }
-
-        /// <summary>
-        /// This is called on a background thread. Do not do synchronous GetService calls here.
-        /// </summary>
-        /// <param name="active"></param>
-        /// <param name="refresh"></param>
-        async void UIContextChanged(bool active, bool refresh)
-        {
-            Debug.Assert(ServiceProvider != null, "UIContextChanged called before service provider is set");
-            if (ServiceProvider == null)
-                return;
-
-            if (active)
-            {
-                if (ActiveRepo == null || refresh)
-                {
-                    ActiveRepo = await System.Threading.Tasks.Task.Run(() =>
-                    {
-                        var repos = GitService.ActiveRepositories;
-                        // Looks like this might return null after a while, for some unknown reason
-                        // if it does, let's refresh the GitService instance in case something got wonky
-                        // and try again. See issue #23
-                        if (repos == null)
-                        {
-                            log.Error("Error 2001: ActiveRepositories is null. GitService: '{GitService}'", GitService);
-
-                            // NOTE: VSGitExt is now self-refreshing
-                            //GitService.Refresh(ServiceProvider);
-
-                            repos = GitService.ActiveRepositories;
-                            if (repos == null)
-                                log.Error("Error 2002: ActiveRepositories is null. GitService: '{GitService}'", GitService);
-                        }
-                        return repos?.FirstOrDefault();
-                    });
-                }
-            }
-            else
-                ActiveRepo = null;
         }
 
         void UpdateActiveRepo()
@@ -229,115 +171,6 @@ namespace GitHub.VisualStudio.Base
         ITeamExplorerPage PageService
         {
             get { return ServiceProvider.GetServiceSafe<ITeamExplorerPage>(); }
-        }
-
-        IVSUIContext GitUIContext
-        {
-            get { return gitUIContext; }
-            set
-            {
-                if (gitUIContext == value)
-                    return;
-                if (gitUIContext != null)
-                    gitUIContext.UIContextChanged -= UIContextChanged;
-                gitUIContext = value;
-                if (gitUIContext != null)
-                    gitUIContext.UIContextChanged += UIContextChanged;
-            }
-        }
-
-        IVSGitExt GitService
-        {
-            get { return gitService; }
-            set
-            {
-                if (gitService == value)
-                    return;
-                if (gitService != null)
-                    gitService.ActiveRepositoriesChanged -= UpdateActiveRepo;
-                gitService = value;
-                if (gitService != null)
-                    gitService.ActiveRepositoriesChanged += UpdateActiveRepo;
-            }
-        }
-
-        IVSUIContextFactory UIContextFactory
-        {
-            get
-            {
-                if (uiContextFactory == null)
-                {
-                    uiContextFactory = ServiceProvider.GetServiceSafe<IVSUIContextFactory>();
-                }
-                return uiContextFactory;
-            }
-        }
-    }
-
-    [Export(typeof(IVSUIContextFactory))]
-    [PartCreationPolicy(CreationPolicy.Shared)]
-    class VSUIContextFactory : IVSUIContextFactory
-    {
-        public IVSUIContext GetUIContext(Guid contextGuid)
-        {
-            return new VSUIContext(UIContext.FromUIContextGuid(contextGuid));
-        }
-    }
-
-    class VSUIContextChangedEventArgs : IVSUIContextChangedEventArgs
-    {
-        public bool Activated { get; }
-
-        public VSUIContextChangedEventArgs(bool activated)
-        {
-            Activated = activated;
-        }
-    }
-
-    class VSUIContext : IVSUIContext
-    {
-        readonly UIContext context;
-        readonly Dictionary<EventHandler<IVSUIContextChangedEventArgs>, EventHandler<UIContextChangedEventArgs>> handlers =
-            new Dictionary<EventHandler<IVSUIContextChangedEventArgs>, EventHandler<UIContextChangedEventArgs>>();
-        public VSUIContext(UIContext context)
-        {
-            this.context = context;
-        }
-
-        public bool IsActive { get { return context.IsActive; } }
-
-        public event EventHandler<IVSUIContextChangedEventArgs> UIContextChanged
-        {
-            add
-            {
-                EventHandler<UIContextChangedEventArgs> handler = null;
-                if (!handlers.TryGetValue(value, out handler))
-                {
-                    handler = (s, e) => value.Invoke(s, new VSUIContextChangedEventArgs(e.Activated));
-                    handlers.Add(value, handler);
-                }
-                context.UIContextChanged += handler;
-            }
-            remove
-            {
-                EventHandler<UIContextChangedEventArgs> handler = null;
-                if (handlers.TryGetValue(value, out handler))
-                {
-                    handlers.Remove(value);
-                    context.UIContextChanged -= handler;
-                }
-            }
-        }
-    }
-
-    static class IGitRepositoryInfoExtensions
-    {
-        /// <summary>
-        /// Create a LocalRepositoryModel from a VS git repo object
-        /// </summary>
-        public static ILocalRepositoryModel ToModel(this IGitRepositoryInfo repo)
-        {
-            return repo == null ? null : new LocalRepositoryModel(repo.RepositoryPath);
         }
     }
 }
