+++ src/GitHub.TeamFoundation.14/Services/VSGitExt.cs
@@ -7,6 +7,7 @@ using GitHub.Services;
 using GitHub.Logging;
 using Serilog;
 using Microsoft.VisualStudio.TeamFoundation.Git.Extensibility;
+using Microsoft.VisualStudio.Shell;
 
 namespace GitHub.VisualStudio.Base
 {
@@ -67,5 +68,72 @@ namespace GitHub.VisualStudio.Base
 
         public IEnumerable<ILocalRepositoryModel> ActiveRepositories => gitService?.ActiveRepositories.Select(x => x.ToModel());
         public event Action ActiveRepositoriesChanged;
+    }
+
+    [Export(typeof(IVSUIContextFactory))]
+    [PartCreationPolicy(CreationPolicy.Shared)]
+    class VSUIContextFactory : IVSUIContextFactory
+    {
+        public IVSUIContext GetUIContext(Guid contextGuid)
+        {
+            return new VSUIContext(UIContext.FromUIContextGuid(contextGuid));
+        }
+    }
+
+    class VSUIContextChangedEventArgs : IVSUIContextChangedEventArgs
+    {
+        public bool Activated { get; }
+
+        public VSUIContextChangedEventArgs(bool activated)
+        {
+            Activated = activated;
+        }
+    }
+
+    class VSUIContext : IVSUIContext
+    {
+        readonly UIContext context;
+        readonly Dictionary<EventHandler<IVSUIContextChangedEventArgs>, EventHandler<UIContextChangedEventArgs>> handlers =
+            new Dictionary<EventHandler<IVSUIContextChangedEventArgs>, EventHandler<UIContextChangedEventArgs>>();
+        public VSUIContext(UIContext context)
+        {
+            this.context = context;
+        }
+
+        public bool IsActive { get { return context.IsActive; } }
+
+        public event EventHandler<IVSUIContextChangedEventArgs> UIContextChanged
+        {
+            add
+            {
+                EventHandler<UIContextChangedEventArgs> handler = null;
+                if (!handlers.TryGetValue(value, out handler))
+                {
+                    handler = (s, e) => value.Invoke(s, new VSUIContextChangedEventArgs(e.Activated));
+                    handlers.Add(value, handler);
+                }
+                context.UIContextChanged += handler;
+            }
+            remove
+            {
+                EventHandler<UIContextChangedEventArgs> handler = null;
+                if (handlers.TryGetValue(value, out handler))
+                {
+                    handlers.Remove(value);
+                    context.UIContextChanged -= handler;
+                }
+            }
+        }
+    }
+
+    static class IGitRepositoryInfoExtensions
+    {
+        /// <summary>
+        /// Create a LocalRepositoryModel from a VS git repo object
+        /// </summary>
+        public static ILocalRepositoryModel ToModel(this IGitRepositoryInfo repo)
+        {
+            return repo == null ? null : new LocalRepositoryModel(repo.RepositoryPath);
+        }
     }
 }
